+++ src/views/DEX/components/OrderBook/OrderBook.tsx
@@ -2,22 +2,26 @@ import React, { memo, useEffect, useMemo, useState } from "react";
 import { Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Typography } from "@mui/material";
 import { CSSProperties } from "@mui/styled-engine";
 import useAPI from "hooks/useAPI";
-import { IGetOrdersResult } from "types/NXTAPI";
+import { IAsset, IGetOrdersResult } from "types/NXTAPI";
 import useBlocks from "hooks/useBlocks";
 import { NQTtoNXT } from "utils/common/NQTtoNXT";
 import { LongUnitPrecision } from "utils/common/constants";
+import { QNTtoNXT } from "utils/common/QNTtoNXT";
+import { BigNumber } from "bignumber.js";
 
 interface IOrderbookProps {
   assetId?: string;
 }
+
 const OrderBook: React.FC<IOrderbookProps> = ({ assetId }) => {
-  const { getOrders } = useAPI();
+  const { getOrders, getAsset } = useAPI();
   const { blockHeight } = useBlocks();
   const [openOrders, setOpenOrders] = useState<IGetOrdersResult>();
+  const [assetDetails, setAssetDetails] = useState<IAsset>();
 
   // maps both bid and ask orders
   const RowsMemo = useMemo(() => {
-    if (openOrders === undefined) {
+    if (openOrders === undefined || assetDetails === undefined || assetDetails === undefined) {
       return;
     }
 
@@ -25,7 +29,7 @@ const OrderBook: React.FC<IOrderbookProps> = ({ assetId }) => {
       return (
         <TableRow key={index}>
           <TableCell>{NQTtoNXT(order.priceNQT, LongUnitPrecision)}</TableCell>
-          <TableCell>{order.quantityQNT}</TableCell>
+          <TableCell>{QNTtoNXT(new BigNumber(order.quantityQNT), assetDetails?.decimals, LongUnitPrecision)}</TableCell>
         </TableRow>
       );
     });
@@ -34,26 +38,30 @@ const OrderBook: React.FC<IOrderbookProps> = ({ assetId }) => {
       return (
         <TableRow key={index}>
           <TableCell>{NQTtoNXT(order.priceNQT, LongUnitPrecision)}</TableCell>
-          <TableCell>{order.quantityQNT}</TableCell>
+          <TableCell>{QNTtoNXT(new BigNumber(order.quantityQNT), assetDetails?.decimals, LongUnitPrecision)}</TableCell>
         </TableRow>
       );
     });
 
     return { asks: mappedAskOrders, bids: mappedBidOrders };
-  }, [openOrders]);
+  }, [assetDetails, openOrders]);
 
   // set the orders for the current asset
   useEffect(() => {
     async function fetchOrders() {
-      if (getOrders === undefined || assetId === undefined) {
+      if (getOrders === undefined || assetId === undefined || getAsset === undefined) {
         return;
       }
 
       setOpenOrders(await getOrders(assetId));
+
+      const assetDetailsResult = await getAsset(assetId); // asset details are required to obtain decimals for proper QNT conversion
+
+      setAssetDetails(assetDetailsResult?.results);
     }
 
     fetchOrders();
-  }, [assetId, blockHeight, getOrders]);
+  }, [assetId, blockHeight, getAsset, getOrders]);
 
   const AskOrderbookMemo = useMemo(() => {
     const askOrderbookStyling: CSSProperties = {