+++ RELEASE-NOTES.md
@@ -17,5 +17,6 @@
 
 ## 0.0.4
 
+- Fix asset precision/quantity issue by properly converting assets with different decimals params
 - Improve text on DEX page when no Asset is selected
 - Add icon and text to address book add button